[FIR] Fix the user project failure
The problem is caused by
`FirExpression.tryToSetSourceForImplicitReceiver()`:
at some point a smartcast expression is
instantiated with an `originalExpression`
with no source, but later it's replaced
with a one that does have a one,
which leads to inconsistency.

Also the receiver expression may have `null` source
instead of a fake source, which can also make
`FirJvmModuleAccessibilityTypeChecker` fail.