IR: copy backing fields when building overrides
Before this commit, fake overrides of field-only properties were
represented as IrPropertyWithLateBinding with all of getter, setter,
backingField equal to null.

Unfortunately this results in a situation where we cannot generate a
proper reference to a field in fir2ir. Basically we need either a getter
or a backing field for each property during fake override construction.
In this commit, this is solved by copying backingField for field-only
properties in IrFakeOverrideBuilder.

However, backend does not support the case when a fake override property
has a backing field, and it leads to all kinds of problems, including
generating additional fields and reporting incorrect conflicting JVM
declarations errors.

So, after all fake overrides are built, we're removing backingField from
all properties created by IrFakeOverrideBuilder. This makes the
resulting IR similar to K1 even in complicated cases, for example, for
property references to fields inherited from Java.